Testing Document
I have started testing the code, there is a document to update when
testing the code which details what is working, what has been tested
etc.
So far the entire Temperature module has been written, and the only
function left to test is readTempx2() which is the one which actually
performs an A/D read so we'll need the hardware to test that. I've also
started writing code and testing for other modules